requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-filter==2.4.0
djangorestframework==3.12.1
idna==2.10
importlib-metadata==2.0.0
Markdown==3.2.2
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
zipp==3.3.0
